Rename class
Đổi tên class từ sv sang mn
