+++ StudentManager/src/Main.java
@@ -1,8 +1,10 @@
-import Service.SystemService;
+import Service.Authentication;
+import Service.SubjectManagement;
 
 public class Main {
     public static void main(String[] args) {
-        SystemService s = new SystemService();
-        s.start();
+        SubjectManagement subjectService = new SubjectManagement();
+        Authentication auth = new Authentication(subjectService);
+        auth.Login();
     }
 }
