+++ StudentManager/src/entities/User.java
@@ -1,18 +1,85 @@
 package entities;
 
 
-public class User extends Infomation{
+public class User{
 
-    private Role role;
-    private boolean isDeleted;
+    protected String code;
+    protected String name;
+    protected String birthDay;
+    protected String gender;
+    protected String address;
+    protected String email;
+    protected String password;
 
-
-
+    protected Role role;
+    protected boolean isDeleted;
 
     public User(String code, String name, String birthDay, String gender, String address, String email, String password, Role role, boolean isDeleted) {
-        super(code, name, birthDay, gender, address, email, password);
+        this.code = code;
+        this.name = name;
+        this.birthDay = birthDay;
+        this.gender = gender;
+        this.address = address;
+        this.email = email;
+        this.password = password;
         this.role = role;
-        this.isDeleted= isDeleted;
+        this.isDeleted = isDeleted;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setBirthDay(String birthDay) {
+        this.birthDay = birthDay;
+    }
+
+    public void setGender(String gender) {
+        this.gender = gender;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getBirthDay() {
+        return birthDay;
+    }
+
+    public String getGender() {
+        return gender;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public String getName() {
+        return name;
     }
 
     public Role getRole(){
@@ -22,6 +89,10 @@ public class User extends Infomation{
     public void setDeleted(boolean deleted)
     {
         isDeleted=deleted;
+    }
+
+    public boolean login(String email, String password) {
+        return this.email.equals(email) && this.password.equals(password);
     }
 
 }
